requirements.txt
Flask==3.0.0
python-dotenv==0.19.0
